+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## alpha-0.0.17
+
+-   Switched default chain to Polygon
+-   Showing address in [EIP-3770](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-3770.md) format.
+
 ## alpha-0.0.16
 
 -   Show multiple transactions log on activity